requirements.txt
# IMPLEMENT: Model-Based Online Policy Iteration for Embodied LLM Agents
# Requirements file

# Core dependencies
python>=3.9,<3.11

# Deep Learning frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.20.0
peft>=0.7.0
trl>=0.7.0

# LLM API
openai>=1.0.0

# Vision and Image Processing
opencv-python>=4.8.0
pillow>=10.0.0

# ALFWorld Environment
alfworld>=0.4.0
ai2thor==2.1.0

# FastAPI for API server
fastapi>=0.100.0
uvicorn>=0.23.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Logging and Monitoring
wandb>=0.15.0

# Optional: for development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

